moved config into processors
