+++ packages/deprecation-crawler/src/lib/processors/config.ts
@@ -1,16 +1,16 @@
-import { CrawlConfig } from './models';
+import { CrawlConfig } from '../models';
 import {
   getConfigPath,
   getVerboseFlag,
   readRepoConfig,
   updateRepoConfig,
-} from './utils';
-import { ensureDeprecationCommentConfig } from './tasks/ensure-deprecation-comment-config';
-import { ensureTsConfigPath } from './tasks/ensure-tsconfig-path';
-import { ensureDeprecationUrlConfig } from './tasks/ensure-deprecations-url-config';
-import { ensureOutputDirectoryConfig } from './tasks/ensure-output-directory-config';
-import { ensureConfigDefaults } from './tasks/ensure-config-defaults';
-import { printHeadline, ProcessFeedback } from './log';
+} from '../utils';
+import { ensureDeprecationCommentConfig } from '../tasks/ensure-deprecation-comment-config';
+import { ensureTsConfigPath } from '../tasks/ensure-tsconfig-path';
+import { ensureDeprecationUrlConfig } from '../tasks/ensure-deprecations-url-config';
+import { ensureOutputDirectoryConfig } from '../tasks/ensure-output-directory-config';
+import { ensureConfigDefaults } from '../tasks/ensure-config-defaults';
+import { printHeadline, ProcessFeedback } from '../log';
 import * as kleur from 'kleur';
 
 const feedback = getSetupFeedback();
